fixed application not running without ivy renderer which causes problems in ie11 (#1825)
