+++ apps/docs/src/app/core/component-docs/alert/examples/alert-component-as-content.module.ts
@@ -11,7 +11,6 @@ import { AlertContentComponent } from './alert-content.component';
 
 @NgModule({
     declarations: [
-        AlertComponentAsContentExampleComponent, AlertContentComponent
     ],
     imports: [
         BrowserModule,
