+++ apps/docs/src/app/core/component-docs/modal/examples/component-as-content/modal-component-as-content-example.module.ts
@@ -14,8 +14,6 @@ import { ModalContentComponent } from './modal-content.component';
 
 @NgModule({
     declarations: [
-        ModalComponentAsContentExampleComponent,
-        ModalContentComponent
     ],
     imports: [
         BrowserModule,
