+++ apps/docs/src/app/core/component-docs/modal/examples/stackable-modals/modal-in-modal-example.module.ts
@@ -13,9 +13,6 @@ import { ModalService, ModalRef } from '@fundamental-ngx/core';
 
 @NgModule({
     declarations: [
-        ModalInModalComponent,
-        ModalInModalSecondComponent,
-        ModalInModalFirstComponent
     ],
     imports: [
         BrowserModule,
