+++ apps/docs/src/app/core/component-docs/tabs/examples/tab-selection-example.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-tab-selection-example',
+    templateUrl: './tab-selection-example.component.html',
+    styleUrls: ['tab-selection-example.component.scss']
+})
+export class TabSelectionExampleComponent {
+    selectedTab = 0;
+}
